+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Core : new `device.nodeJS` to detect if running under node.js
 - Particle Emitter: added the possibility to define a tint for particles
+- Renderer: added "additive" as an alias for the "lighter" blend mode
 
 ### Changed
 - Core : replace internal use of global window object by `globalThis`