+++ CHANGELOG.md
@@ -7,12 +7,14 @@
 - Particle Emitter: added the possibility to define a tint for particles
 - Renderer: added "additive" as an alias for the "lighter" blend mode
 
+
 ### Changed
 - Core : replace internal use of global window object by `globalThis`
 - Core : refactor the boot process, and device/feature detection/initialisation to better work within node.js (will still require jsdom and node-canvas)
 
 ### Fixed
 - Particle Emitter: fix a regression in `ParticleEmitter` causing a crash, after es6 class reformatting
+- Particle Emitter: fix particle additive blend mode in WebGL
 
 
 ## [10.6.1] (melonJS 2) - _2022-04-25_